# packages/shared/openapi.yaml
openapi: 3.1.0
info:
  title: Professor-for-a-Day Product API
  version: 0.1.0
  description: |
    Contract-first API for the Professor-for-a-Day MVP.

    This document is the authoritative agreement between the web frontend and the
    FastAPI backend (see docs/adr/0001-openapi-is-the-product-api-contract.md).
    The backend is verified against this file; generated output never redefines it.

    Scope: only the product routes the web app consumes. The legacy `/api/chat`
    and `/api/conversations` routes are intentionally excluded.

    Key design decisions:
    - Voice input is two-step: `POST /api/speech/transcriptions` turns audio into
      text, then the transcript is submitted through the ordinary text-turn
      contract with `input_mode: "voice"`. Transcription failures never touch a
      session.
    - Voice output is synthesize-on-fetch: JSON responses carry no audio. The
      client fetches `GET /api/sessions/{session_id}/turns/{turn_number}/speech`
      when it wants a reply spoken (see docs/adr/0003-speech-is-synthesized-on-fetch.md).
      Clients should cache fetched audio blobs for replay.
    - Errors use a minimal envelope `{"error": {"code", "message"}}` with an
      enumerated, provider-neutral `code`. Vendors, models, upstream statuses,
      and upstream error text never appear in a response.
    - Turn submissions are idempotent via a client-generated `client_turn_id`.
    - Mastery is browser-local and appears nowhere in this contract.
...

...
security: []
servers:
...
paths:
...
  /api/graphs:
    get:
      tags: [graphs]
      operationId: listGraphs
      summary: List knowledge graphs
      description: |
        Returns every knowledge graph: the builtin Machine Learning graph first,
        then user-created graphs oldest-first. No LLM call, no rubric content,
        no per-learner state.
      responses:
        "200":
          description: All knowledge graphs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphList"
              example:
                graphs:
                  - { id: machine-learning, title: Machine Learning, source: builtin, concept_count: 15, created_at: null }
                  - { id: 66bcd1f2a9c4e35d8f01a2b3, title: Compilers, source: user, concept_count: 6, created_at: "2026-08-15T09:30:00Z" }
        "503":
          $ref: "#/components/responses/DatabaseUnavailable"
  /api/graphs/{graph_id}:
    delete:
      tags: [graphs]
      operationId: deleteGraph
      summary: Delete a user-created knowledge graph
      description: |
        Permanently removes a user-created graph (its concepts, edges, and
        cached generated rubrics). The builtin `machine-learning` graph can
        never be deleted. Past Teaching Sessions and their reports are the
        learning record and are never touched by graph deletion.
      parameters:
        - $ref: "#/components/parameters/GraphId"
      responses:
        "204":
          description: The graph was deleted.
        "404":
          description: No such knowledge graph.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: { code: GRAPH_NOT_FOUND, message: "No such knowledge graph." }
        "409":
          description: The builtin graph cannot be deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: { code: GRAPH_NOT_DELETABLE, message: "The builtin graph cannot be deleted." }
        "503":
          $ref: "#/components/responses/DatabaseUnavailable"
  /api/graphs/{graph_id}/curriculum:
    get:
      tags: [graphs]
      operationId: getGraphCurriculum
      summary: One graph's concepts and prerequisite edges
      description: |
        Returns the graph's Concepts and the directed prerequisite edges between
        them. The builtin graph is backend-owned, version-controlled data; user
        graphs are stored server-side and grow append-only at session end. No
        LLM call, no rubric content, no per-learner state. Edges are
        recommendations, never locks.
      parameters:
        - $ref: "#/components/parameters/GraphId"
      responses:
        "200":
          description: The graph's full curriculum.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Curriculum"
              example:
                concepts:
                  - { id: dataset, title: Dataset, summary: "A collection of examples used to teach or evaluate a model." }
                  - { id: features-and-labels, title: Features and Labels, summary: "The inputs a model sees and the answers it learns to predict." }
                  - { id: model, title: Model, summary: "A function with learnable parameters that maps features to predictions." }
                  - { id: training-vs-inference, title: Training vs. Inference, summary: "Learning parameters from data versus using them to predict." }
                  - { id: supervised-learning, title: Supervised Learning, summary: "Learning from examples that include the correct answer." }
                  - { id: unsupervised-learning, title: Unsupervised Learning, summary: "Finding structure in data without labeled answers." }
                  - { id: neural-networks, title: Neural Networks, summary: "Layered models that learn hierarchical representations." }
                  - { id: loss-function, title: Loss Function, summary: "A number that measures how wrong the model's predictions are." }
                  - { id: gradient-descent, title: Gradient Descent, summary: "Iteratively adjusting parameters downhill along the loss gradient." }
                  - { id: learning-rate, title: Learning Rate, summary: "The step size taken on each gradient descent update." }
                  - { id: overfitting, title: Overfitting, summary: "Memorizing training data at the cost of generalization." }
                  - { id: regularization, title: Regularization, summary: "Techniques that penalize complexity to improve generalization." }
                  - { id: train-validation-test-split, title: Train/Validation/Test Split, summary: "Separating data to train, tune, and honestly evaluate." }
                  - { id: confusion-matrix, title: Confusion Matrix, summary: "A table of prediction outcomes against true classes." }
                  - { id: precision-vs-recall, title: Precision vs. Recall, summary: "Two complementary views of a classifier's mistakes." }
                edges:
                  - { from: dataset, to: features-and-labels }
                  - { from: features-and-labels, to: model }
                  - { from: model, to: training-vs-inference }
                  - { from: training-vs-inference, to: supervised-learning }
                  - { from: training-vs-inference, to: unsupervised-learning }
                  - { from: supervised-learning, to: neural-networks }
                  - { from: model, to: loss-function }
                  - { from: loss-function, to: gradient-descent }
                  - { from: gradient-descent, to: learning-rate }
                  - { from: supervised-learning, to: overfitting }
                  - { from: overfitting, to: regularization }
                  - { from: overfitting, to: train-validation-test-split }
                  - { from: supervised-learning, to: confusion-matrix }
                  - { from: confusion-matrix, to: precision-vs-recall }
        "404":
          description: No such knowledge graph.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: { code: GRAPH_NOT_FOUND, message: "No such knowledge graph." }
        "503":
          $ref: "#/components/responses/DatabaseUnavailable"
...
components:
  parameters:
    GraphId:
      name: graph_id
      in: path
      required: true
      description: The knowledge-graph identifier from `GET /api/graphs`.
      schema:
        $ref: "#/components/schemas/GraphId"
...
  responses:
...
    DatabaseUnavailable:
      description: The database is not available.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error: { code: DB_UNAVAILABLE, message: "The database is not available." }
  schemas:
...
    GraphId:
      type: string
      pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
      description: Stable graph identifier. `machine-learning` is the builtin graph; user graphs use server-generated ids.
    GraphSource:
      type: string
      enum: [builtin, user]
    GraphSummary:
      type: object
      required: [id, title, source, concept_count, created_at]
      properties:
        id:
          $ref: "#/components/schemas/GraphId"
        title:
          type: string
        source:
          $ref: "#/components/schemas/GraphSource"
        concept_count:
          type: integer
          minimum: 1
        created_at:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          description: Null for the builtin graph, which predates the database.
    GraphList:
      type: object
      required: [graphs]
      properties:
        graphs:
          type: array
          items:
            $ref: "#/components/schemas/GraphSummary"
...
    Curriculum:
      type: object
      required: [concepts, edges]
      properties:
        concepts:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Concept"
        edges:
          type: array
          description: Directed prerequisite recommendations. Acyclic; never locks a node.
          items:
            $ref: "#/components/schemas/PrerequisiteEdge"
    Concept:
      type: object
      required: [id, title, summary]
      properties:
        id:
          $ref: "#/components/schemas/ConceptId"
        title:
          type: string
        summary:
          type: string
          description: Short learner-facing description. Never contains rubric content.
    ConceptId:
      type: string
      pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
      description: Stable, URL-safe concept identifier (e.g. `gradient-descent`).
...
    PrerequisiteEdge:
      type: object
      required: [from, to]
      properties:
        from:
          $ref: "#/components/schemas/ConceptId"
        to:
          $ref: "#/components/schemas/ConceptId"
...
    ErrorCode:
      type: string
      enum:
        - INVALID_GRAPH
        - INVALID_CONCEPT
        - INVALID_MODE
        - EMPTY_SUBMISSION
        - GRAPH_NOT_FOUND
        - GRAPH_NOT_DELETABLE
        - SESSION_NOT_FOUND
        - TURN_NOT_FOUND
        - SESSION_ENDED
        - TRANSCRIPTION_FAILED
        - GENERATION_FAILED
        - SPEECH_FAILED
        - UPLOAD_TOO_LARGE
        - UNSUPPORTED_AUDIO_TYPE
        - DB_UNAVAILABLE
        - VALIDATION_FAILED
      description: |
        Provider-neutral error codes. `VALIDATION_FAILED` is the fallback for
        request-shape errors not covered by a more specific code.
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
              description: Human-readable, provider-neutral. Never names a vendor, model, upstream status, or upstream error text.